temporarily move to phi-3

+++ download_models.sh
@@ -26,6 +26,11 @@ set -e
 
   for file_name in "${file_names[@]}"
   do
+    # check if file exists
+    if [ -f "phi-3-mini-128k-chat/$file_name" ]; then
+      echo "phi-3-mini-128k-chat/$file_name exists, skipping..."
+      continue
+    fi
     wget -O phi-3-mini-128k-chat/$file_name https://huggingface.co/microsoft/Phi-3-mini-128k-instruct-onnx/resolve/main/cpu_and_mobile/cpu-int4-rtn-block-32-acc-level-4/$file_name
   done
 )
